Fix databricks build break
