+++ iceberg/iceberg-stub/src/main/scala/com/nvidia/spark/rapids/iceberg/IcebergProbeImpl.scala
@@ -16,8 +16,6 @@
 
 package com.nvidia.spark.rapids.iceberg
 
-import com.nvidia.spark.rapids.ShimReflectionUtils
-
 /**
  * Stub IcebergProbe for Spark versions where Iceberg is not supported.
  * Similar to DeltaProbeImpl in delta-stub.
@@ -26,9 +24,8 @@ class IcebergProbeImpl extends IcebergProbe {
   override def getDetectedVersion: String =
     throw new UnsupportedOperationException("Iceberg is not supported in this configuration")
 
-  override def shimPackage: String = "com.nvidia.spark.rapids.iceberg.icebergstub"
+  override def shimPackage: String =
+    throw new UnsupportedOperationException("Iceberg is not supported in this configuration")
 
-  override def getProvider: IcebergProvider =
-    ShimReflectionUtils.newInstanceOf[IcebergProvider](
-      s"${shimPackage}.IcebergProviderImpl")
+  override def getProvider: IcebergProvider = NoIcebergProvider
 }
