--- iceberg/iceberg-stub/src/main/scala/com/nvidia/spark/rapids/iceberg/icebergstub/IcebergProviderImpl.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2026, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nvidia.spark.rapids.iceberg.icebergstub
-
-import com.nvidia.spark.rapids.{GpuExec, GpuExpression, ScanRule, SparkPlanMeta, StaticInvokeMeta}
-import com.nvidia.spark.rapids.iceberg.IcebergProvider
-
-import org.apache.spark.sql.catalyst.expressions.objects.StaticInvoke
-import org.apache.spark.sql.connector.read.Scan
-import org.apache.spark.sql.connector.write.Write
-import org.apache.spark.sql.execution.SparkPlan
-
-class IcebergProviderImpl extends IcebergProvider {
-  private def unsupported: Nothing =
-    throw new UnsupportedOperationException("Iceberg is not supported in this configuration")
-
-  override def getScans: Map[Class[_ <: Scan], ScanRule[_ <: Scan]] = Map.empty
-  override def tagForGpu(expr: StaticInvoke, meta: StaticInvokeMeta): Unit = unsupported
-  override def convertToGpu(expr: StaticInvoke, meta: StaticInvokeMeta): GpuExpression = unsupported
-  override def isSupportedWrite(write: Class[_ <: Write]): Boolean = false
-  override def isSupportedCatalog(catalogClass: Class[_]): Boolean = false
-  override def tagForGpuPlan[P <: SparkPlan, M <: SparkPlanMeta[P]](
-      cpuExec: P, meta: M): Unit = unsupported
-  override def convertToGpuPlan[P <: SparkPlan, M <: SparkPlanMeta[P]](
-      cpuExec: P, meta: M): GpuExec = unsupported
-}
