+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/iceberg/IcebergProvider.scala
@@ -77,3 +77,18 @@ object IcebergProvider {
     }
   }
 }
+
+object NoIcebergProvider extends IcebergProvider {
+  private def unsupported: Nothing =
+    throw new UnsupportedOperationException("Iceberg is not supported in this configuration")
+
+  override def getScans: Map[Class[_ <: Scan], ScanRule[_ <: Scan]] = Map.empty
+  override def tagForGpu(expr: StaticInvoke, meta: StaticInvokeMeta): Unit = unsupported
+  override def convertToGpu(expr: StaticInvoke, meta: StaticInvokeMeta): GpuExpression = unsupported
+  override def isSupportedWrite(write: Class[_ <: Write]): Boolean = false
+  override def isSupportedCatalog(catalogClass: Class[_]): Boolean = false
+  override def tagForGpuPlan[P <: SparkPlan, M <: SparkPlanMeta[P]](
+      cpuExec: P, meta: M): Unit = unsupported
+  override def convertToGpuPlan[P <: SparkPlan, M <: SparkPlanMeta[P]](
+      cpuExec: P, meta: M): GpuExec = unsupported
+}
